NETFLIX LANDING PAGE
Welcome to my project showcasing a reimagined Netflix landing page, brought to life with the magic of HTML, CSS, and JavaScript. As an avid fan of both web development and streaming content, I embarked on this project to merge my passions and create a dynamic user experience.

The HTML structure forms the backbone, ensuring a well-organized and accessible layout. CSS swoops in to stylize each element, capturing the sleek and modern aesthetic synonymous with Netflix. From the bold typography to the responsive design, every detail has been meticulously crafted to mirror the streaming giant's signature look.

JavaScript takes the user interaction to the next level. Enjoy a carousel of featured shows and movies that dynamically update, providing a taste of the diverse content available. Immerse yourself in the interactive navigation bar, offering seamless transitions and a user-friendly experience.